LeRoy Dataset - k-TSP code for classification of metabolites from PTM histones
Implementation of k-Top Scoring Pairs algorithm for classification of 136 metabolites using PTM histone data. This implementation is for the LeRoy dataset. The model is trained on CCLE Dataset and tested on the LeRoy dataset. Model evaluation is done by calculating performance parameters like accuracy, MCC, precision, recall and F1-score for each of the metabolite.